add key usage editing
This adds the ability to edit key usage flags in --edit-key. Enter
"usage" to perform an edit. If subkeys are selected, each will be
edited in turn; otherwise the primary will be edited.

This code does insist that (per spec) a primary key retains the
certification flag, and key usage flags are still limited to the
capabilities of the key's algo. Aside from that you can make any
changes you like.